add a theme constants file
