+++ app/features/theme/index.ts
@@ -1,0 +1,4 @@
+export { useTheme } from './use-theme';
+export { ThemeProvider } from './theme-provider';
+export * from './theme-cookies.server';
+export type * from './theme.types';
